[CARBONDATA-3289] MV datamap doesn't take effect when having clause use alias
